[chore] : refactor domain test packages
