+++ src/test/java/nextstep/courses/domain/image/ImagePixelTest.java
@@ -1,6 +1,5 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
-import nextstep.courses.domain.image.ImagePixel;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
