+++ src/test/java/nextstep/courses/domain/image/ImageTest.java
@@ -1,9 +1,5 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
-import nextstep.courses.domain.image.Image;
-import nextstep.courses.domain.image.ImagePixel;
-import nextstep.courses.domain.image.ImageSize;
-import nextstep.courses.domain.image.ImageType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
