+++ src/test/java/nextstep/courses/domain/session/FreeSessionTest.java
@@ -1,11 +1,9 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.domain.image.Image;
 import nextstep.courses.domain.image.ImagePixel;
 import nextstep.courses.domain.image.ImageSize;
 import nextstep.courses.domain.image.ImageType;
-import nextstep.courses.domain.session.FreeSession;
-import nextstep.courses.domain.session.SessionDate;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
