+++ src/test/java/nextstep/courses/service/CourseServiceTest.java
@@ -1,12 +1,12 @@
 package nextstep.courses.service;
 
-import nextstep.courses.domain.LecturerTest;
 import nextstep.courses.domain.course.Course;
 import nextstep.courses.domain.course.CourseRepository;
 import nextstep.courses.domain.image.Image;
 import nextstep.courses.domain.image.ImagePixel;
 import nextstep.courses.domain.image.ImageSize;
 import nextstep.courses.domain.image.ImageType;
+import nextstep.courses.domain.lecturer.LecturerTest;
 import nextstep.courses.domain.session.FreeSession;
 import nextstep.courses.domain.session.SessionDate;
 import nextstep.users.domain.NsUserTest;
